move inside worker tests to subfolder
