+++ tests/conversion/converters/inside_worker_test/declarative_schema/osm_models.py
