+++ tests/conversion/converters/inside_worker_test/dumps/pbf/monaco/monaco-latest.osm.pbf
