+++ tests/conversion/converters/inside_worker_test/dumps/sql/osm.schema.sql.bz2
