+++ tests/conversion/converters/inside_worker_test/label_sql_test.py
@@ -4,8 +4,8 @@ import pytest
 import sqlalchemy
 
 from tests.conftest import TagCombination
-from tests.inside_worker_test.conftest import sql_from_bootstrap_relative_location, slow
-from tests.inside_worker_test.declarative_schema import osm_models
+from tests.conversion.converters.inside_worker_test.conftest import sql_from_bootstrap_relative_location, slow
+from tests.conversion.converters.inside_worker_test.declarative_schema import osm_models
 
 xeno = "大洲南部広域農道"
 xeno_transliterated = 'dà zhōu nán bù guǎng yù nóng dào'
